+++ harmoni-data/src/main/java/org/harmoniapp/harmonidata/repositories/ShiftRepository.java
@@ -3,5 +3,5 @@ package org.harmoniapp.harmonidata.repositories;
 import org.harmoniapp.harmonidata.entities.Shift;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShiftRepository extends JpaRepository<Shift, Integer> {
+public interface ShiftRepository extends JpaRepository<Shift, Long> {
 }
