+++ harmoni-data/src/main/java/org/harmoniapp/harmonidata/repositories/VacationRepository.java
@@ -3,5 +3,5 @@ package org.harmoniapp.harmonidata.repositories;
 import org.harmoniapp.harmonidata.entities.Vacation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VacationRepository extends JpaRepository<Vacation, Integer> {
+public interface VacationRepository extends JpaRepository<Vacation, Long> {
 }
